Migrated interface module and update definition conventions
