+++ PVGeo/_helpers/__init__.py
@@ -1,8 +1,8 @@
-from .readers import *
 from .arrays import *
-from .timeseries import *
-from .xml import *
 from .errors import *
+from .timeseries import *
+from .readers import *
+from .xml import *
 
 
 __displayname__ = 'Internal Helpers'
