+++ PVGeo/filters/voxelize.py
@@ -12,6 +12,7 @@ from ..base import FilterBase
 from .. import _helpers
 from ..version import checkNumpy
 from .xyz import RotationTool
+from .. import interface
 
 ###############################################################################
 
@@ -58,7 +59,7 @@ class VoxelizePoints(FilterBase):
     def AddCellData(grid, arr, name):
         """Add a NumPy array as cell data to the given grid input
         """
-        c = _helpers.numToVTK(arr)
+        c = interface.numToVTK(arr)
         c.SetName(name)
         grid.GetCellData().AddArray(c)
         return grid
@@ -155,7 +156,7 @@ class VoxelizePoints(FilterBase):
             self.AddFieldData(grid)
 
         # Add unique nodes as points in output
-        pts.SetData(_helpers.numToVTK(unique_nodes))
+        pts.SetData(interface.numToVTK(unique_nodes))
 
         # Add cell vertices
         j = np.multiply(np.tile(np.arange(0, 8, 1), numCells), numCells)
