+++ PVGeo/grids/extract_topo.py
@@ -9,6 +9,7 @@ from datetime import datetime
 # Import Helpers:
 from ..base import FilterBase
 from .. import _helpers
+from .. import interface
 # NOTE: internal import - from scipy.spatial import cKDTree
 
 
@@ -85,7 +86,7 @@ class ExtractTopography(FilterBase):
         active = np.array(datapts[:,2] < (comp[:,2] - self.__tolerance), dtype=int)
 
         # Now add cell data to output
-        active = _helpers.numToVTK(active)
+        active = interface.numToVTK(active)
         active.SetName('Active Topography')
         grid.GetCellData().AddArray(active)
         return 1
