+++ PVGeo/ubc/octree.py
@@ -17,6 +17,7 @@ else:
 from ..base import AlgorithmBase
 from .two_file_base import ubcMeshReaderBase, ModelAppenderBase
 from .. import _helpers
+from .. import interface
 
 
 
@@ -165,7 +166,7 @@ class OcTreeReader(ubcMeshReaderBase):
             vec_full_nx[unique_nodes['f0']].reshape(-1, 1),
             vec_full_ny[unique_nodes['f1']].reshape(-1, 1),
             vec_full_nz[unique_nodes['f2']].reshape(-1, 1)), axis=1)
-        vtkPtsData = _helpers.numToVTK(ptsArr, deep=1)
+        vtkPtsData = interface.numToVTK(ptsArr, deep=1)
         vtkPts = vtk.vtkPoints()
         vtkPts.SetData(vtkPtsData)
 
@@ -184,7 +185,7 @@ class OcTreeReader(ubcMeshReaderBase):
         pdo.SetCells(vtk.VTK_VOXEL, CellArr)
 
         # Add the indexing of the cell's
-        vtkIndexArr = _helpers.numToVTK(ind_cell_corner.ravel(), deep=1)
+        vtkIndexArr = interface.numToVTK(ind_cell_corner.ravel(), deep=1)
         vtkIndexArr.SetName('index_cell_corner')
         pdo.GetCellData().AddArray(vtkIndexArr)
 
@@ -221,7 +222,7 @@ class OcTreeReader(ubcMeshReaderBase):
         model = model[ind_reorder]
 
         # Convert data to VTK data structure and append to output
-        c = _helpers.numToVTK(model, deep=True)
+        c = interface.numToVTK(model, deep=True)
         c.SetName(dataNm)
         # THIS IS CELL DATA! Add the model data to CELL data:
         mesh.GetCellData().AddArray(c)
@@ -259,7 +260,7 @@ class OcTreeReader(ubcMeshReaderBase):
         # Get output:
         output = self.GetOutputData(outInfo, 0)
         # Get requested time index
-        i = _helpers.GetRequestedTime(self, outInfo)
+        i = _helpers.getRequestedTime(self, outInfo)
         self.__ubcOcTree(
             self.GetMeshFileName(),
             self.GetModelFileNames(),
